don't add the encrypt and decrypt methods to the model
