+++ attr_secure.gemspec
@@ -23,6 +23,5 @@ Gem::Specification.new do |spec|
   spec.add_development_dependency "minitest", "~> 5.0.3"
 
   spec.add_dependency 'fernet'
-  spec.add_dependency 'activesupport', ' ~> 3.2.0'
   spec.add_dependency 'activerecord',  ' ~> 3.2.0'
 end
