+++ lib/attr_secure.rb
@@ -2,42 +2,23 @@ require "attr_secure/version"
 require 'fernet'
 require 'active_record'
 
+require 'attr_secure/secure'
+
 Fernet::Configuration.run do |config|
   config.enforce_ttl = false
 end
 
 module AttrSecure
-  extend ActiveSupport::Concern
 
-  def env!(key)
-    ENV[key] || raise("Missing ENV(#{key})")
-  end
-
-  def encrypt(value)
-    Fernet.generate(env!('ATTR_SECURE_SECRET')) do |generator|
-      generator.data = { value: value }
-    end
-  end
-
-  def decrypt(value)
-    return nil if value.nil?
-    verifier = Fernet.verifier(env!('ATTR_SECURE_SECRET'), value)
-    verifier.data["value"] if verifier.valid?
-  end
-
-  module ClassMethods
-
-    def attr_secure(attribute)
-      define_method("#{attribute}=") do |value|
-        write_attribute(attribute, encrypt(value.nil? ? nil : value.to_sym))
-      end
-
-      define_method("#{attribute}") do
-        decrypt read_attribute(attribute.to_sym)
-      end
+  def attr_secure(attribute)
+    define_method("#{attribute}=") do |value|
+      write_attribute(attribute, Secure.new.encrypt(value.nil? ? nil : value.to_sym))
     end
 
+    define_method("#{attribute}") do
+      Secure.new.decrypt read_attribute(attribute.to_sym)
+    end
   end
 end
 
-ActiveRecord::Base.send(:include, AttrSecure)
+ActiveRecord::Base.send(:extend, AttrSecure)
